+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Change Log
 4.7.0-beta.3 (unreleased)
 --------------------------
 
+### Improvements
+
+- [beta.2] Adjust the layout of the Quick Action bar.
+
 
 
 4.7.0-beta.2 (613)