+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Change Log
 
 ### Fixes
 
+- Fixed an issue that the “unescape replacement string” option could not unescape consecutive backslashes correctly.
 - [beta] Fix an issue that some active actions were not listed in the Quick Action bar.