+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Change Log
 
 - Enable the “Input Backslash” and “Input Yen Mark” menu commands in all text fields in the application.
 - [trivial] Add a stepper to the custom tab width sheet.
-- [beta] Use the same shortcut symbols as the menu display in the Quick Action bar.  
+- [beta] Use the same shortcut symbols as the menu display in the Quick Action bar.
 - [beta] Adjust the layout of the Quick Action bar.
 - [beta.2] Update some localized text.
 
@@ -18,7 +18,7 @@ Change Log
 
 - Fixed an issue that the “unescape replacement string” option could not unescape consecutive backslashes correctly.
 - [beta] Fix an issue that some active actions were not listed in the Quick Action bar.
-- [beta] Fix an issue that keyboard shortcuts that include the Globe/fn key were not displayed in the Quick Action.
+- [beta] Fix an issue that keyboard shortcuts that include the Globe/fn key were not displayed in the Quick Action view.